Refactor snarkvm-fields
