+++ algorithms/src/commitment/pedersen_parameters.rs
@@ -21,7 +21,7 @@ use crate::{
     traits::CRH,
 };
 use snarkvm_curves::traits::Group;
-use snarkvm_fields::{errors::ConstraintFieldError, traits::ToConstraintField, Field};
+use snarkvm_fields::{ConstraintFieldError, Field, ToConstraintField};
 use snarkvm_utilities::bytes::{FromBytes, ToBytes};
 
 use rand::Rng;
