+++ algorithms/src/crh/pedersen.rs
@@ -22,7 +22,7 @@ use crate::{
     traits::{CRHParameters, CRH},
 };
 use snarkvm_curves::Group;
-use snarkvm_fields::{errors::ConstraintFieldError, traits::to_field_vec::ToConstraintField, Field};
+use snarkvm_fields::{ConstraintFieldError, Field, ToConstraintField};
 use snarkvm_utilities::bytes_to_bits;
 
 use rand::Rng;
