+++ curves/src/bls12_377/fq12.rs
@@ -14,9 +14,10 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::bls12_377::{Fq, Fq2, Fq6Parameters};
 use snarkvm_fields::{field, Fp12, Fp12Parameters};
 use snarkvm_utilities::biginteger::BigInteger384 as BigInteger;
+
+use crate::bls12_377::{Fq, Fq2, Fq6Parameters};
 
 pub type Fq12 = Fp12<Fq12Parameters>;
