+++ curves/src/bls12_377/fq6.rs
@@ -14,13 +14,14 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::bls12_377::{Fq, Fq2, Fq2Parameters};
 use snarkvm_fields::{
     field,
     fp6_3over2::{Fp6, Fp6Parameters},
     Fp2Parameters,
 };
 use snarkvm_utilities::biginteger::BigInteger384;
+
+use crate::bls12_377::{Fq, Fq2, Fq2Parameters};
 
 pub type Fq6 = Fp6<Fq6Parameters>;
 
@@ -254,12 +255,13 @@ impl Fp6Parameters for Fq6Parameters {
 
 #[cfg(test)]
 mod test {
-    use super::*;
+    use rand::SeedableRng;
+    use rand_xorshift::XorShiftRng;
+
     use snarkvm_fields::{One, Zero};
     use snarkvm_utilities::rand::UniformRand;
 
-    use rand::SeedableRng;
-    use rand_xorshift::XorShiftRng;
+    use super::*;
 
     #[test]
     fn test_fq2_mul_nonresidue() {
