+++ curves/src/bw6_761/fq6.rs
@@ -14,12 +14,13 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::bw6_761::{Fq, Fq3, Fq3Parameters};
 use snarkvm_fields::{
     field,
     fp6_2over3::{Fp6, Fp6Parameters},
 };
 use snarkvm_utilities::biginteger::BigInteger768 as BigInteger;
+
+use crate::bw6_761::{Fq, Fq3, Fq3Parameters};
 
 pub type Fq6 = Fp6<Fq6Parameters>;
