+++ curves/src/bw6_761/g2.rs
@@ -14,12 +14,13 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
+use snarkvm_fields::{field, Zero};
+use snarkvm_utilities::biginteger::{BigInteger384, BigInteger768};
+
 use crate::{
     bw6_761::{Fq, Fr},
     traits::{ModelParameters, SWModelParameters},
 };
-use snarkvm_fields::{field, Zero};
-use snarkvm_utilities::biginteger::{BigInteger384, BigInteger768};
 
 #[derive(Clone, Default, PartialEq, Eq)]
 pub struct Bls12_377G2Parameters;
