+++ curves/src/templates/bls12/bls12.rs
@@ -23,11 +23,12 @@ use crate::{
 };
 use serde::{Deserialize, Serialize};
 use snarkvm_fields::{
-    fp12_2over3over2::{Fp12, Fp12Parameters},
-    fp2::Fp2Parameters,
     fp6_3over2::Fp6Parameters,
     Field,
+    Fp12,
+    Fp12Parameters,
     Fp2,
+    Fp2Parameters,
     One,
     PrimeField,
     SquareRootField,
