+++ dpc/src/base_dpc/program/program.rs
@@ -16,7 +16,7 @@
 
 use crate::base_dpc::BaseDPCComponents;
 use snarkvm_algorithms::traits::{CommitmentScheme, CRH};
-use snarkvm_fields::{errors::ConstraintFieldError, traits::to_field_vec::ToConstraintField};
+use snarkvm_fields::{ConstraintFieldError, ToConstraintField};
 
 /// Program verification key and proof
 /// Represented as bytes to be generic for any Program SNARK
