+++ gadgets/src/algorithms/snark/gm17.rs
@@ -28,7 +28,7 @@ use crate::traits::{
 };
 use snarkvm_algorithms::snark::gm17::{Proof, VerifyingKey, GM17};
 use snarkvm_curves::traits::{AffineCurve, PairingEngine};
-use snarkvm_fields::{traits::to_field_vec::ToConstraintField, Field};
+use snarkvm_fields::{Field, ToConstraintField};
 use snarkvm_r1cs::{errors::SynthesisError, ConstraintSynthesizer, ConstraintSystem};
 use snarkvm_utilities::bytes::FromBytes;
