+++ gadgets/src/fields/fp12.rs
@@ -28,9 +28,10 @@ use crate::{
     },
 };
 use snarkvm_fields::{
-    fp12_2over3over2::{Fp12, Fp12Parameters},
     fp6_3over2::{Fp6, Fp6Parameters},
     Field,
+    Fp12,
+    Fp12Parameters,
     Fp2Parameters,
     One,
     PrimeField,
