+++ gadgets/src/traits/utilities/uint/unsigned_integer.rs
@@ -14,6 +14,13 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
+use core::borrow::Borrow;
+use std::{cmp::Ordering, fmt::Debug};
+
+use snarkvm_fields::{Field, FpParameters, PrimeField, ToConstraintField};
+use snarkvm_r1cs::{errors::SynthesisError, Assignment, ConstraintSystem, LinearCombination};
+use snarkvm_utilities::bytes::ToBytes;
+
 use crate::{
     fields::FpGadget,
     utilities::{
@@ -25,13 +32,6 @@ use crate::{
         ToBytesGadget,
     },
 };
-use snarkvm_fields::{fp_parameters::FpParameters, traits::to_field_vec::ToConstraintField, Field, PrimeField};
-use snarkvm_r1cs::{errors::SynthesisError, Assignment, ConstraintSystem, LinearCombination};
-
-use snarkvm_utilities::bytes::ToBytes;
-
-use core::borrow::Borrow;
-use std::{cmp::Ordering, fmt::Debug};
 
 uint_impl!(UInt8, u8, 8);
 uint_impl!(UInt16, u16, 16);
